variables cleaning

+++ src/help.cpp
@@ -82,7 +82,7 @@ namespace
 		}
 		if (result.length() > 12)
 			result = result.substr(0, 12);
-		for (int i = result.length(); i <= 12; result += " ", i++) { }
+		for (size_t i = result.length(); i <= 12; result += " ", i++) { }
 		result += ": ";
 		return result;
 	}
